portfolio_analyzer/requirements.txt
streamlit
PyGithub
spacy
sentence-transformers
scikit-learn
pandas
numpy
umap-learn
openai
beautifulsoup4
Jinja2
xhtml2pdf
plotly
nltk
python-dotenv 